chore: move traits for safe and finalized block read/write
